Add files via upload
Predicting the likelihood of the person having a bank account or not (Yes = 1, No = 0), for each unique id in the test dataset . I trained my model on 80% of the data and test your model on the final 20% of the data, across four East African countries - Kenya, Rwanda, Tanzania, and Uganda. I evaluated my model using the Mean Squared Error.